Move excel export to `src/exports`.
